+++ pyluos/modules/encoder.py
