+++ tests/test_dxl.py
@@ -2,7 +2,7 @@ import unittest
 
 from contextlib import closing
 
-from pyrobus import Robot
+from pyluos import Robot
 
 import fakerobot
