+++ tests/test_import.py
@@ -3,10 +3,10 @@ import unittest
 
 class TestImport(unittest.TestCase):
     def test_import_general(self):
-        import pyrobus
+        import pyluos
 
     def test_import_robot(self):
-        from pyrobus import Robot
+        from pyluos import Robot
 
 
 if __name__ == '__main__':
