+++ tests/test_servo.py
@@ -3,7 +3,7 @@ import unittest
 from threading import Event
 from contextlib import closing
 
-from pyrobus import Robot
+from pyluos import Robot
 
 import fakerobot
