+++ tox.ini
@@ -6,4 +6,4 @@ deps =
   nose
   coverage
   tornado
-commands = nosetests --with-coverage --cover-erase --cover-package=pyrobus -v -w tests/
+commands = nosetests --with-coverage --cover-erase --cover-package=pyluos -v -w tests/
